rename features folder
