+++ frontend/src/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Counter from './features/counter/Counter';
+import Counter from './components/counter/Counter';
 import './App.css';
 
 function App() {
