+++ frontend/src/components/exampleInput/exampleInput.stories.tsx
